Updated controllers fn
