+++ controllers/contacts/getById.js
@@ -1,0 +1,25 @@
+const contactsOperations = require("../../models/contacts");
+
+const getById = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const contact = await contactsOperations.getContactById(contactId);
+    if (!contact) {
+      res.status(404).json({
+        status: "error",
+        code: 404,
+        message: "Not found",
+      });
+      return;
+    }
+    res.json({
+      status: "success",
+      code: 200,
+      data: { result: contact },
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = getById;
