+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+const getAll = require("./getAll");
+const getById = require("./getById");
+const add = require("./add");
+const updateById = require("./updateByid");
+const removeById = require("./removeById");
+
+module.exports = {
+  getAll,
+  getById,
+  add,
+  updateById,
+  removeById,
+};
