+++ controllers/contacts/removeById.js
@@ -1,0 +1,25 @@
+const contactsOperations = require("../../models/contacts");
+
+const removeById = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const removedContact = await contactsOperations.removeContact(contactId);
+    if (!removedContact) {
+      res.status(404).json({
+        status: "error",
+        code: 404,
+        message: "Not found",
+      });
+      return;
+    }
+    res.json({
+      status: "success",
+      code: 200,
+      message: "contact deleted",
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = removeById;
