+++ controllers/contacts/updateByid.js
@@ -1,0 +1,44 @@
+const contactsOperations = require("../../models/contacts");
+
+const Joi = require("joi");
+const contactSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().email().required(),
+  phone: Joi.string().required(),
+});
+
+const updateById = async (req, res, next) => {
+  try {
+    const { error } = contactSchema.validate(req.body);
+    if (error) {
+      res.status(400).json({
+        status: "error",
+        code: 400,
+        message: "missing fields",
+      });
+      return;
+    }
+    const { contactId } = req.params;
+    const updatedContact = await contactsOperations.updateContact(
+      contactId,
+      req.body
+    );
+    if (!updatedContact) {
+      res.status(404).json({
+        status: "error",
+        code: 404,
+        message: "Not found",
+      });
+      return;
+    }
+    res.json({
+      status: "success",
+      code: 200,
+      data: { result: updatedContact },
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = updateById;
