+++ routes/api/contacts.js
@@ -1,127 +1,16 @@
 const express = require("express");
-
 const router = express.Router();
 
-const Joi = require("joi");
-const contactSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().email().required(),
-  phone: Joi.string().required(),
-});
+const { contacts: ctrl } = require("../../controllers");
 
-const contactsOperations = require("../../models/contacts");
+router.get("/", ctrl.getAll);
 
-router.get("/", async (req, res, next) => {
-  try {
-    const contacts = await contactsOperations.listContacts();
-    res.json({
-      status: "success",
-      code: 200,
-      data: { result: contacts },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/:contactId", ctrl.getById);
 
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const contact = await contactsOperations.getContactById(contactId);
-    if (!contact) {
-      res.status(404).json({
-        status: "error",
-        code: 404,
-        message: "Not found",
-      });
-      return;
-    }
-    res.json({
-      status: "success",
-      code: 200,
-      data: { result: contact },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.post("/", ctrl.add);
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = contactSchema.validate(req.body);
-    if (error) {
-      res.status(400).json({
-        status: "error",
-        code: 400,
-        message: "missing required name field",
-      });
-      return;
-    }
-    const newContact = await contactsOperations.addContact(req.body);
-    res.status(201).json({
-      status: "success",
-      code: 201,
-      data: { result: newContact },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete("/:contactId", ctrl.removeById);
 
-router.delete("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const removedContact = await contactsOperations.removeContact(contactId);
-    if (!removedContact) {
-      res.status(404).json({
-        status: "error",
-        code: 404,
-        message: "Not found",
-      });
-      return;
-    }
-    res.json({
-      status: "success",
-      code: 200,
-      message: "contact deleted",
-    });
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    const { error } = contactSchema.validate(req.body);
-    if (error) {
-      res.status(400).json({
-        status: "error",
-        code: 400,
-        message: "missing fields",
-      });
-      return;
-    }
-    const { contactId } = req.params;
-    const updatedContact = await contactsOperations.updateContact(
-      contactId,
-      req.body
-    );
-    if (!updatedContact) {
-      res.status(404).json({
-        status: "error",
-        code: 404,
-        message: "Not found",
-      });
-      return;
-    }
-    res.json({
-      status: "success",
-      code: 200,
-      data: { result: updatedContact },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.put("/:contactId", ctrl.updateById);
 
 module.exports = router;
